% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BJstatistics.R
\name{BJStat}
\alias{BJStat}
\title{Compute Berk-Jones statistics}
\usage{
BJStat(x, indexL = seq_along(x), indexU = seq_along(x))
}
\arguments{
\item{x}{data}

\item{indexL}{lower side index}

\item{indexU}{upper side index}
}
\description{
Compute Berk-Jones statistics
}
